refactor: reorganize billing service structure
