+++ tenant-platform/billing-service/src/main/java/com/ejada/billing/config/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Configuration components for the billing service.
+ */
+package com.ejada.billing.config;
